Separate vscode code from parsing code for python
